Adding controllers

+++ CoffeeShop/CoffeeShop.Core/Abstract/Patterns/Payment.cs
@@ -2,6 +2,6 @@
 {
     public interface Payment
     {
-        void Pay(decimal price);
+        void Pay(double price);
     }
 }
