+++ CoffeeShop/CoffeeShop.Core/Patterns/Strategy/PayByCash.cs
@@ -4,7 +4,7 @@ namespace CoffeeShop.Core.Patterns.Strategy
 {
     public class PayByCash : Payment
     {
-        public void Pay(decimal price)
+        public void Pay(double price)
         {
             Console.WriteLine($"I will pay by cash {price} lei");
         }
